+++ src/main/java/sesac/server/account/controller/AccountController.java
@@ -11,11 +11,11 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
-import sesac.server.account.dto.EmailCheckRequest;
-import sesac.server.account.dto.LoginRequest;
-import sesac.server.account.dto.LoginResponse;
-import sesac.server.account.dto.LogoutRequest;
-import sesac.server.account.dto.SignupRequest;
+import sesac.server.account.dto.request.EmailCheckRequest;
+import sesac.server.account.dto.request.LoginRequest;
+import sesac.server.account.dto.response.LoginResponse;
+import sesac.server.account.dto.request.LogoutRequest;
+import sesac.server.account.dto.request.SignupRequest;
 import sesac.server.account.exception.AccountErrorCode;
 import sesac.server.account.service.AccountService;
 import sesac.server.auth.dto.AuthPrincipal;
