+++ src/main/java/sesac/server/account/dto/request/LogoutRequest.java
@@ -1,4 +1,4 @@
-package sesac.server.account.dto;
+package sesac.server.account.dto.request;
 
 public record LogoutRequest(String accessToken, String refreshToken) {
