+++ src/main/java/sesac/server/account/service/AccountService.java
@@ -9,9 +9,9 @@ import lombok.RequiredArgsConstructor;
 import lombok.extern.log4j.Log4j2;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
-import sesac.server.account.dto.LoginRequest;
-import sesac.server.account.dto.LoginResponse;
-import sesac.server.account.dto.SignupRequest;
+import sesac.server.account.dto.request.LoginRequest;
+import sesac.server.account.dto.response.LoginResponse;
+import sesac.server.account.dto.request.SignupRequest;
 import sesac.server.account.exception.AccountErrorCode;
 import sesac.server.account.exception.AccountException;
 import sesac.server.campus.entity.Course;
