+++ src/main/java/sesac/server/feed/controller/CampusFeedController.java
@@ -19,11 +19,11 @@ import org.springframework.web.bind.annotation.RestController;
 import sesac.server.auth.dto.AuthPrincipal;
 import sesac.server.auth.dto.CustomPrincipal;
 import sesac.server.common.exception.BindingResultHandler;
-import sesac.server.feed.dto.CreatePostRequest;
-import sesac.server.feed.dto.PostListRequest;
-import sesac.server.feed.dto.PostListResponse;
-import sesac.server.feed.dto.PostResponse;
-import sesac.server.feed.dto.UpdatePostRequest;
+import sesac.server.feed.dto.request.CreatePostRequest;
+import sesac.server.feed.dto.request.PostListRequest;
+import sesac.server.feed.dto.response.PostListResponse;
+import sesac.server.feed.dto.response.PostResponse;
+import sesac.server.feed.dto.request.UpdatePostRequest;
 import sesac.server.feed.entity.PostType;
 import sesac.server.feed.exception.PostErrorCode;
 import sesac.server.feed.service.PostService;
