+++ src/main/java/sesac/server/feed/dto/request/CreatePostRequest.java
@@ -1,4 +1,4 @@
-package sesac.server.feed.dto;
+package sesac.server.feed.dto.request;
 
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.Size;
