+++ src/main/java/sesac/server/feed/dto/request/PostListRequest.java
@@ -1,4 +1,4 @@
-package sesac.server.feed.dto;
+package sesac.server.feed.dto.request;
 
 import sesac.server.feed.entity.PostType;
