+++ src/main/java/sesac/server/feed/dto/request/UpdatePostRequest.java
@@ -1,4 +1,4 @@
-package sesac.server.feed.dto;
+package sesac.server.feed.dto.request;
 
 import jakarta.validation.constraints.Size;
