+++ src/main/java/sesac/server/feed/repository/search/PostSearch.java
@@ -3,8 +3,8 @@ package sesac.server.feed.repository.search;
 import java.util.List;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
-import sesac.server.feed.dto.PostListRequest;
-import sesac.server.feed.dto.PostListResponse;
+import sesac.server.feed.dto.request.PostListRequest;
+import sesac.server.feed.dto.response.PostListResponse;
 import sesac.server.feed.entity.PostType;
 
 public interface PostSearch {
