+++ src/main/java/sesac/server/feed/service/PostService.java
@@ -10,12 +10,12 @@ import sesac.server.auth.dto.CustomPrincipal;
 import sesac.server.auth.exception.TokenErrorCode;
 import sesac.server.auth.exception.TokenException;
 import sesac.server.common.exception.BaseException;
-import sesac.server.feed.dto.CreatePostRequest;
-import sesac.server.feed.dto.PostListRequest;
-import sesac.server.feed.dto.PostListResponse;
-import sesac.server.feed.dto.PostResponse;
-import sesac.server.feed.dto.ReplyResponse;
-import sesac.server.feed.dto.UpdatePostRequest;
+import sesac.server.feed.dto.request.CreatePostRequest;
+import sesac.server.feed.dto.request.PostListRequest;
+import sesac.server.feed.dto.response.PostListResponse;
+import sesac.server.feed.dto.response.PostResponse;
+import sesac.server.feed.dto.response.ReplyResponse;
+import sesac.server.feed.dto.request.UpdatePostRequest;
 import sesac.server.feed.entity.FeedType;
 import sesac.server.feed.entity.Hashtag;
 import sesac.server.feed.entity.Post;
