+++ src/main/java/sesac/server/manager/service/ManagerPostService.java
@@ -7,10 +7,10 @@ import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 import sesac.server.common.exception.BaseException;
-import sesac.server.feed.dto.PostListRequest;
-import sesac.server.feed.dto.PostListResponse;
-import sesac.server.feed.dto.PostResponse;
-import sesac.server.feed.dto.ReplyResponse;
+import sesac.server.feed.dto.request.PostListRequest;
+import sesac.server.feed.dto.response.PostListResponse;
+import sesac.server.feed.dto.response.PostResponse;
+import sesac.server.feed.dto.response.ReplyResponse;
 import sesac.server.feed.entity.Post;
 import sesac.server.feed.entity.PostType;
 import sesac.server.feed.exception.PostErrorCode;
